Performed complete website redesign to integrate content in single-page format
After a brand new design test that turned out to be successful (See w3css_cards_float.html), I have removed the redundant multi-page component of the project and performed a complete overhaul to the front-end component of the website. All data is now integrated into a single webpage for user-convenience as well as design factor & ease of maintenance.